first approximation: store network in script
